fix(Constant): prize를 int에서 long으로 수정

+++ src/main/java/lotto/domain/constants/Constants.java
@@ -10,12 +10,12 @@ public class Constants {
 
     public static final int LOTTO_PRICE = 1000;
 
-    public static final int FIRST_PRIZE = 2_000_000_000;
-    public static final int SECOND_PRIZE = 30_000_000;
-    public static final int THIRD_PRIZE = 1_500_000;
-    public static final int FOURTH_PRIZE = 50_000;
-    public static final int FIFTH_PRIZE = 5_000;
-    public static final int NO_PRIZE = 0;
+    public static final long FIRST_PRIZE = 2_000_000_000;
+    public static final long SECOND_PRIZE = 30_000_000;
+    public static final long THIRD_PRIZE = 1_500_000;
+    public static final long FOURTH_PRIZE = 50_000;
+    public static final long FIFTH_PRIZE = 5_000;
+    public static final long NO_PRIZE = 0;
 
     // 객체 생성 방지
     private Constants() {}
